api_yamdb/requirements.txt
requests==2.26.0
django-filter==0.9.2
django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
